refactor to make package and class names more suitable
