+++ src/test/java/uk/gov/ons/ctp/integration/fieldsvccucumber/RunCucumberTestSSO.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.contcencucumber;
+package uk.gov.ons.ctp.integration.fieldsvccucumber;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
@@ -8,6 +8,6 @@ import org.junit.runner.RunWith;
 @CucumberOptions(
     plugin = {"pretty", "html:target/cucumber"},
     features = {"src/test/resources/integrationtests/sso"},
-    glue = {"uk.gov.ons.ctp.integration.contcencucumber.cucSteps.sso"},
+    glue = {"uk.gov.ons.ctp.integration.fieldsvccucumber.cucSteps.sso"},
     dryRun = false)
 public class RunCucumberTestSSO {}
