+++ src/test/java/uk/gov/ons/ctp/integration/fieldsvccucumber/selenium/pageobject/SSO.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.contcencucumber.selenium.pageobject;
+package uk.gov.ons.ctp.integration.fieldsvccucumber.selenium.pageobject;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
